pyproject.toml
[project]
name = "ecoprimers"
version = "1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.4.1",
    "flask>=3.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.1.0",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.0",
    "newsapi-python>=0.2.7",
    "openmeteo-requests>=1.3.0",
    "openmeteo-sdk>=1.18.6",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "sendgrid>=6.11.0",
    "sqlalchemy-libsql>=0.1.0",
    "stripe>=11.5.0",
]